add ft_ping beta

+++ ft_ping/src/main.c
@@ -6,7 +6,7 @@
 /*   By: mbougrin <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/10/18 11:02:44 by mbougrin          #+#    #+#             */
-/*   Updated: 2016/10/26 11:51:18 by mbougrin         ###   ########.fr       */
+/*   Updated: 2016/10/26 11:51:54 by mbougrin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -227,8 +227,8 @@ void					printSigint(void)
 		ret--;	
 	}
 	printf("\n--- %s %s statistics ---\n", stc->ip, stc->name);
-	printf("%d packets transmitted, %d received, %d\% packet loss, time %fms", \
-			stc->count, stc->packetReceiv, (int)ret, stc->allMs);
+	printf("%d packets transmitted, %d received, %f\% packet loss, time %fms", \
+			stc->count, stc->packetReceiv, ret, stc->allMs);
 	free(stc);
 	exit(0);
 }
